minor update #71
