+++ srcs/execution/exec_pipe_list.c
@@ -14,31 +14,30 @@
 
 extern t_request	g_request;
 
-t_bool	is_enable_environ_path(void)
+void	call_execve_function(const char **cmd_args)
 {
-	t_environ *path;
+	char		**environs;
+	struct stat	buf;
 
-	path = get_target_environ("PATH");
-	if (!path || path->value[0] == '\0')
-		return (FALSE);
-	return (TRUE);
-}
-
-static void replace_path(char **cmd, char *path)
-{
-	free(cmd[0]);
-	cmd[0] = ft_strdup(path);
+	environs = env_list_to_array(g_request.environs);
+	if (execve(cmd_args[0], (char *const *)cmd_args, environs) < 0)
+	{
+		stat(cmd_args[0], &buf);
+		if (!access(cmd_args[0], R_OK))
+			exit(0);
+		else if (!buf.st_size)
+			print_err_and_exit(ERR_MSG_PERM_DENIED, DENIED);
+		else
+			print_err_and_exit(NULL, GNRL_ERR);
+	}
 }
 
 void	exec_path_cmd(t_pipe_list *pipe_list)
 {
 	const char	**cmd_args;
-	char		**environs;
 	t_exit_cd	exit_cd;
-	struct stat	buf;
 	char		*err_msg;
 
-	environs = env_list_to_array(g_request.environs);
 	cmd_args = pipe_list->cmd_args;
 	err_msg = NULL;
 	exit_cd = SCCSS;
@@ -48,62 +47,18 @@ void	exec_path_cmd(t_pipe_list *pipe_list)
 		if (exit_cd != SCCSS)
 			print_err_and_exit_free(&err_msg, exit_cd);
 		free(err_msg);
-		if (execve(cmd_args[0], (char *const *)cmd_args, environs) < 0)
-		{
-			stat(cmd_args[0], &buf);
-			if (!access(cmd_args[0], R_OK))
-				exit(0);
-			else if (!buf.st_size)
-				print_err_and_exit(ERR_MSG_PERM_DENIED, DENIED);
-			else
-				print_err_and_exit(NULL, GNRL_ERR);
-		}
+		call_execve_function(cmd_args);
 	}
 	else
 	{
 		if (is_enable_environ_path() == FALSE)
 			print_err_and_exit(ERR_MSG_NO_FILE, CMD_NOT_FND);
-
-		t_environ *path;
-		char **paths;
-		int i;
-		t_bool	flag;
-
-		path = get_target_environ("PATH");
-		paths = split_path(path->value, ':');
-		i = -1;
-		flag = FALSE;
-		while (paths[++i])
-		{
-			exit_cd = SCCSS;
-			paths[i] = add_slash(paths[i]);
-			paths[i] = join_path(paths[i], (char *)cmd_args[0]);
-			exit_cd = check_executable_cmd_path(paths[i], NULL);
-			if (!flag && exit_cd == DENIED)
-				flag = TRUE;
-			if (exit_cd != SCCSS)
-				continue;
-			replace_path((char **)cmd_args, paths[i]);
-			break;
-		}
-		multi_free(paths);
-		if (exit_cd != SCCSS && flag)
-			exit_cd = DENIED;
+		exit_cd = search_path((char **)cmd_args);
 		if (exit_cd == DENIED)
 			print_err_and_exit(ERR_MSG_PERM_DENIED, DENIED);
 		else if (exit_cd != SCCSS)
 			print_err_and_exit(ERR_MSG_NO_FILE, CMD_NOT_FND);
-
-		if (execve(cmd_args[0], (char *const *)cmd_args, environs) < 0)
-		{
-			stat(cmd_args[0], &buf);
-			if (!access(cmd_args[0], R_OK))
-				exit(0);
-			else if (!buf.st_size)
-				print_err_and_exit(ERR_MSG_PERM_DENIED, DENIED);
-			else
-				print_err_and_exit(NULL, GNRL_ERR);
-		}
+		call_execve_function(cmd_args);
 	}
 }
 
@@ -120,26 +75,6 @@ void	child_exec_cmd(t_pipe_list *pipe_list)
 		g_request.builtin_funcs[builtin_id](pipe_list->cmd_args, TRUE);
 	else
 		exec_path_cmd(pipe_list);
-}
-
-void	wait_processes(t_pipe_list *pipe_list)
-{
-	pid_t		changed_pid;
-	t_pipe_list	*tmp_node;
-	int			status;
-
-	tmp_node = pipe_list;
-	while (tmp_node)
-	{
-		changed_pid = waitpid(tmp_node->pid, &status, 0);
-		g_request.exit_cd = WEXITSTATUS(status);
-		if (changed_pid < 0)
-		{
-			perror("waitpid");
-			return ;
-		}
-		tmp_node = tmp_node->next;
-	}
 }
 
 void	execute_cmds(t_pipe_list *pipe_list)
