+++ srcs/execution/serch_cmd_path.c
@@ -6,7 +6,7 @@
 /*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/18 11:32:56 by jnakahod          #+#    #+#             */
-/*   Updated: 2021/10/25 23:31:22 by jnakahod         ###   ########.fr       */
+/*   Updated: 2021/10/30 16:41:34 by jnakahod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,25 +18,32 @@ static void	replace_path(char **cmd, char *path)
 	cmd[0] = ft_strdup(path);
 }
 
-t_bool	search_path(char **cmd)
+t_exit_cd	search_path(char **cmd)
 {
 	t_environ	*path;
 	char		**paths;
 	int			i;
+	t_bool		deniedable;
+	t_exit_cd	exit_cd;
 
 	path = get_target_environ("PATH");
 	paths = split_path(path->value, ':');
 	i = -1;
+	deniedable = FALSE;
 	while (paths[++i])
 	{
 		paths[i] = add_slash(paths[i]);
 		paths[i] = join_path(paths[i], cmd[0]);
-		if (access(paths[i], X_OK) == -1)
-			continue ;
-		replace_path(cmd, paths[i]);
-		multi_free(paths);
-		return (TRUE);
+		exit_cd = check_executable_cmd_path(paths[i], NULL);
+		if (!deniedable && exit_cd == DENIED)
+			deniedable = TRUE;
+		if (exit_cd != SCCSS)
+			continue;
+		replace_path((char **)cmd, paths[i]);
+		break;
 	}
 	multi_free(paths);
-	return (FALSE);
+	if (exit_cd != SCCSS && deniedable)
+		exit_cd = DENIED;
+	return (exit_cd);
 }
