+++ srcs/utils/utils3.c
@@ -6,7 +6,7 @@
 /*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/16 12:25:40 by jnakahod          #+#    #+#             */
-/*   Updated: 2021/10/19 22:53:05 by jnakahod         ###   ########.fr       */
+/*   Updated: 2021/10/30 16:31:56 by jnakahod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,4 +40,14 @@ t_bool	has_pipe(t_pipe_list *pipe_list)
 	if (pipe_list->next)
 		return (TRUE);
 	return (FALSE);
+}
+
+t_bool	is_enable_environ_path(void)
+{
+	t_environ *path;
+
+	path = get_target_environ("PATH");
+	if (!path || path->value[0] == '\0')
+		return (FALSE);
+	return (TRUE);
 }
